# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e5533a452a8e9561a0d0793
  title: Security
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /search:
    post:
      x-skaffolder-id: 5e554ff152a8e9561a0d08ea
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Search
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Distance:
                  type: number
                Place:
                  type: string
    get:
      x-skaffolder-id: 5e554fe552a8e9561a0d08e2
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Search
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Search"
  "/search/{id}":
    delete:
      x-skaffolder-id: 5e554ff152a8e9561a0d08f2
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Search
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e554ff152a8e9561a0d08ec
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Search
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
    post:
      x-skaffolder-id: 5e554ff152a8e9561a0d08eb
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Search
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Search"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Search"
  "/search/findBySearch/{key}":
    get:
      x-skaffolder-id: 5e554fe552a8e9561a0d08e4
      x-skaffolder-name: findBySearch
      x-skaffolder-id-resource: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-resource: Search
      x-skaffolder-crudAction: findBySearch
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBySearch
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBySearch/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBySearch
      tags:
        - Search
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Search"
...
  /security:
    post:
      x-skaffolder-id: 5e554fe552a8e9561a0d08df
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Security
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Security"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Distance:
                  type: number
                Name:
                  type: string
                Place:
                  type: string
    get:
      x-skaffolder-id: 5e554fe552a8e9561a0d08e3
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Security
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Security"
  "/security/{id}":
    delete:
      x-skaffolder-id: 5e554fe552a8e9561a0d08e7
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Security
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e554fb252a8e9561a0d08dd
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Security
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Security
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Security"
    post:
      x-skaffolder-id: 5e554fe552a8e9561a0d08e0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Security
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Security"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Security"
  "/security/findBySecurity/{key}":
    get:
      x-skaffolder-id: 5e554ff152a8e9561a0d08ef
      x-skaffolder-name: findBySecurity
      x-skaffolder-id-resource: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-resource: Security
      x-skaffolder-crudAction: findBySecurity
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBySecurity
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBySecurity/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBySecurity
      tags:
        - Security
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Security"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Search:
      x-skaffolder-id: 5e554e9d52a8e9561a0d08d8
      x-skaffolder-id-db: 5e5533a452a8e9561a0d0794
      x-skaffolder-id-entity: 5e554e9d52a8e9561a0d08d5
      x-skaffolder-url: /search
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Distance:
          type: number
          x-skaffolder-id-attr: 5e554e9d52a8e9561a0d08d7
          x-skaffolder-type: Number
          x-skaffolder-required: true
        Place:
          type: string
          x-skaffolder-id-attr: 5e554e9d52a8e9561a0d08d6
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        Search:
          x-skaffolder-id: 5e554f7752a8e9561a0d08dc
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e554e9d52a8e9561a0d08d5
          x-skaffolder-ent2: 5e554e5e52a8e9561a0d08cd
      required:
        - _id
        - Distance
        - Place
    Security:
      x-skaffolder-id: 5e554e5e52a8e9561a0d08d1
      x-skaffolder-id-db: 5e5533a452a8e9561a0d0794
      x-skaffolder-id-entity: 5e554e5e52a8e9561a0d08cd
      x-skaffolder-url: /security
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Distance:
          type: number
          x-skaffolder-id-attr: 5e554e5e52a8e9561a0d08d0
          x-skaffolder-type: Number
        Name:
          type: string
          x-skaffolder-id-attr: 5e554e5e52a8e9561a0d08ce
          x-skaffolder-type: String
        Place:
          type: string
          x-skaffolder-id-attr: 5e554e5e52a8e9561a0d08cf
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        Search:
          x-skaffolder-id: 5e554f5552a8e9561a0d08db
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e554e5e52a8e9561a0d08cd
          x-skaffolder-ent2: 5e554e5e52a8e9561a0d08cd
        Security:
          x-skaffolder-id: 5e554f5552a8e9561a0d08da
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5e554e5e52a8e9561a0d08cd
          x-skaffolder-ent2: 5e554e9d52a8e9561a0d08d5
      required:
        - _id
        - Place
...